Use UK spelling of Analyse
