+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/CommandLine/AccountCommandLine.cs
@@ -1,10 +1,10 @@
 ﻿using CommandLine;
 
-namespace SFA.DAS.EAS.LevyAnalyzer.CommandLine
+namespace SFA.DAS.EAS.LevyAnalyser.CommandLine
 {
     public class AccountCommandLine
     {
-        [Option('a', "accounts", HelpText = "Account to analyze (in print page format, e.g. 1-5,8,11)")]
+        [Option('a', "accounts", HelpText = "Account to analyse (in print page format, e.g. 1-5,8,11)")]
         public string AccountIds { get; set; }
     }
 }
