+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/CommandLine/AnalyzeCommandLine.cs
@@ -1,14 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using CommandLine;
+﻿using CommandLine;
 
-namespace SFA.DAS.EAS.LevyAnalyzer.CommandLine
+namespace SFA.DAS.EAS.LevyAnalyser.CommandLine
 {
-    [Verb("aggregate", HelpText = "Run an aggregate query")]
-    public class AnalyzeCommandLine : AccountCommandLine
+    [Verb("analyse", HelpText = "Run an aggregate query")]
+    public class AnalyseCommandLine : AccountCommandLine
     {
         [Option('r', "rules", HelpText = "List of rules to run", Required = false)]
         public string Rules { get; set; }
