+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/Config/ConnectionStringEntry.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.EAS.LevyAnalyzer.Interfaces
+﻿namespace SFA.DAS.EAS.LevyAnalyser.Config
 {
     public class ConnectionStringEntry
     {
