+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/Interfaces/ICommand.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 
-namespace SFA.DAS.EAS.LevyAnalyzer.Interfaces
+namespace SFA.DAS.EAS.LevyAnalyser.Interfaces
 {
     public interface ICommand
     {
