+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/Interfaces/IConfigProvider.cs
@@ -1,5 +1,8 @@
-﻿namespace SFA.DAS.EAS.LevyAnalyzer.Interfaces
+﻿namespace SFA.DAS.EAS.LevyAnalyser.Interfaces
 {
+    /// <summary>
+    ///     Returns the specified config type. 
+    /// </summary>
     public interface IConfigProvider
     {
         TConfigType Get<TConfigType>() where TConfigType : class, new();
