+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/Interfaces/IRuleEvaluationResult.cs
@@ -1,13 +1,21 @@
 ﻿using System.Collections.Generic;
-using SFA.DAS.EAS.LevyAnalyzer.Rules;
-using SFA.DAS.EAS.LevyAnalyzer.Rules.Infrastructure;
+using SFA.DAS.EAS.LevyAnalyser.Rules.Infrastructure;
 
-namespace SFA.DAS.EAS.LevyAnalyzer.Interfaces
+namespace SFA.DAS.EAS.LevyAnalyser.Interfaces
 {
+    /// <summary>
+    ///     Represents the result of applying a single rule to a single account.
+    /// </summary>
     public interface IRuleEvaluationResult
     {
         string RuleName { get; }
+
+        /// <summary>
+        ///     Indicates whether the account obeys the rule. This will be false if 
+        ///     <see cref="Messages"/> contains any <see cref="RuleEntryLevel.Violation"/> messages.
+        /// </summary>
         bool IsValid { get; }
+
         IReadOnlyCollection<RuleEvaluationResultEntry> Messages { get; }
     }
 }
