+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/IoC/FinanceRegistry.cs
@@ -1,7 +1,7 @@
 ﻿using SFA.DAS.EmployerFinance.Services;
 using StructureMap;
 
-namespace SFA.DAS.EAS.LevyAnalyzer.IoC
+namespace SFA.DAS.EAS.LevyAnalyser.IoC
 {
     public class FinanceRegistry : Registry
     {
