+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/IoC/IoC.cs
@@ -1,6 +1,6 @@
 ﻿using StructureMap;
 
-namespace SFA.DAS.EAS.LevyAnalyzer.IoC
+namespace SFA.DAS.EAS.LevyAnalyser.IoC
 {
     public static class IoC
     {
