+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/Models/LevyDeclaration.cs
@@ -1,7 +1,6 @@
 ﻿using System;
-using System.ComponentModel.DataAnnotations;
 
-namespace SFA.DAS.EAS.LevyAnalyzer.Models
+namespace SFA.DAS.EAS.LevyAnalyser.Models
 {
     public class LevyDeclaration
     {
