+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/Models/TransactionLine.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace SFA.DAS.EAS.LevyAnalyzer.Models
+namespace SFA.DAS.EAS.LevyAnalyser.Models
 {
     public enum TransactionItemType : byte
     {
