--- src/Utils/SFA.DAS.EAS.LevyAnalyzer/Repositories/IDbContextFactory.cs
@@ -1,7 +1,0 @@
-﻿namespace SFA.DAS.EAS.LevyAnalyzer.Repositories
-{
-    public interface IDbContextFactory
-    {
-        FinanceDbContext GetFinanceDbContext();
-    }
-}
