+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/Rules/Infrastructure/RuleEntryLevel.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.EAS.LevyAnalyzer.Rules.Infrastructure
+﻿namespace SFA.DAS.EAS.LevyAnalyser.Rules.Infrastructure
 {
     public enum RuleEntryLevel
     {
