+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/Rules/Infrastructure/RuleEvaluationResultEntry.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.EAS.LevyAnalyzer.Rules.Infrastructure
+﻿namespace SFA.DAS.EAS.LevyAnalyser.Rules.Infrastructure
 {
     public class RuleEvaluationResultEntry
     {
