+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/Rules/OntimeLevyShouldResultInATransaction.cs
@@ -1,10 +1,10 @@
 ﻿using System.Linq;
-using SFA.DAS.EAS.LevyAnalyzer.Interfaces;
-using SFA.DAS.EAS.LevyAnalyzer.Models;
-using SFA.DAS.EAS.LevyAnalyzer.Rules.Infrastructure;
+using SFA.DAS.EAS.LevyAnalyser.Interfaces;
+using SFA.DAS.EAS.LevyAnalyser.Models;
+using SFA.DAS.EAS.LevyAnalyser.Rules.Infrastructure;
 using SFA.DAS.EmployerFinance.Services;
 
-namespace SFA.DAS.EAS.LevyAnalyzer.Rules
+namespace SFA.DAS.EAS.LevyAnalyser.Rules
 {
     public class OntimeLevyShouldResultInATransaction : IRule
     {
