+++ src/Utils/SFA.DAS.EAS.LevyAnalyzer/SFA.DAS.EAS.LevyAnalyser.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{F0D2AB9A-D11D-40BE-9B80-A3FDC42AE9F4}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>SFA.DAS.EAS.LevyAnalyzer</RootNamespace>
-    <AssemblyName>SFA.DAS.EAS.LevyAnalyzer</AssemblyName>
+    <RootNamespace>SFA.DAS.EAS.LevyAnalyser</RootNamespace>
+    <AssemblyName>SFA.DAS.EAS.LevyAnalyser</AssemblyName>
     <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -80,9 +80,9 @@
   <ItemGroup>
     <Compile Include="CommandLine\AccountCommandLine.cs" />
     <Compile Include="CommandLine\AnalyzeCommandLine.cs" />
-    <Compile Include="Interfaces\ConnectionStringEntry.cs" />
-    <Compile Include="Interfaces\DatabaseConfig.cs" />
-    <Compile Include="Interfaces\DbConnectionFactory.cs" />
+    <Compile Include="Config\ConnectionStringEntry.cs" />
+    <Compile Include="Config\DatabaseConfig.cs" />
+    <Compile Include="Repositories\DbConnectionFactory.cs" />
     <Compile Include="Interfaces\IRule.cs" />
     <Compile Include="Interfaces\IRuleEvaluationResult.cs" />
     <Compile Include="Interfaces\IRuleRepository.cs" />
@@ -114,7 +114,7 @@
     <Compile Include="Repositories\AccountRepository.cs" />
     <Compile Include="Repositories\DbContextFactory.cs" />
     <Compile Include="Repositories\FinanceDbContext.cs" />
-    <Compile Include="Repositories\IDbContextFactory.cs" />
+    <Compile Include="Interfaces\IDbContextFactory.cs" />
     <Compile Include="ResultSavers\FileResultSavers.cs" />
   </ItemGroup>
   <ItemGroup>
